Add comment on Waffle chart scale computation
Sadly, there's no way to access the value of an interactive selector
in `altair`. Because of this, the color scaling of the github-like
waffle chart might crap up. This is still better than having the plot
almost illegible due to very faint color scales (anchoring the upper
bound to the max number of commits in a given day).

More on this here: https://stackoverflow.com/questions/68329301/fix-scale-botttom-colour-on-0-in-altair